# optimize dns propagation request