Currently the DNS propagation request has a heavy and unnecessary bottleneck. Under the hood the route53 waiter stays in a thread blocking loop checking a single entry every 30 seconds over 60 attempts. Check out the [source code here](https://github.com/aws/aws-sdk-go/blob/v1.37.15/service/route53/waiters.go#L28). The current implementation makes a single change, then waits for the propagation to finish before making the second change. AWS averages around 60 seconds for propagation, which means the current implementation takes about (60s)n, increasing by 60-90 seconds for each entry that needs to be changed, potentially creating very long wait times

This is a simple optimization extracted from a previous PR #22360 and discussed with @fjl to make all of the DNS changes first, and only after, wait for the response for each one. This improves the time complexity to roughly (60s)1. If I had 3 changes to make, the original implementation would take around 3-4 minutes. The new implementation bundles the changes together so the operation should now take ~60-90 seconds total, and would not increase dramatically with more changes